kafkaPython/requirements.txt
aenum==2.2.6
aiohttp==3.7.4
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.5
chardet==3.0.4
click==8.0.1
cryptography==3.4.7
decorator==4.4.2
docker==5.0.0
english==2020.7.0
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
gremlinpython==3.5.0
gunicorn==20.1.0
idna==2.10
isodate==0.6.0
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==2.0.2
MarkupSafe==2.0.1
monotonic==1.6
multidict==5.1.0
neotime==1.7.4
nest-asyncio==1.5.1
networkx==2.5.1
packaging==21.0
pansi==2020.7.3
prompt-toolkit==3.0.19
py2neo==2021.1.5
pycparser==2.20
Pygments==2.9.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.20
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
websocket-client==1.1.0
Werkzeug==2.0.1
yarl==1.6.3
